+++ R/modify_cells2.R
@@ -28,3 +28,19 @@ px_valuenotex <- function(x, value) {
 px_valuenotex.px <- function(x, value) {
   handle_cells(x, value, "2", "valuenotex")
 }
+
+
+#' @rdname px_values.px
+#' @export
+px_values <- function(x, value) {
+  UseMethod("px_values")
+}
+
+#' @inherit px_precision.px
+#' @title VALUES
+#' @description `r table_description("VALUES")`
+#' @param value `r cells_param_value("VALUES", "2")`
+#' @export
+px_values.px <- function(x, value) {
+  handle_cells(x, value, "2", "value")
+}