+++ tests/testthat/test-55-modify-cells2.R
@@ -1,4 +1,4 @@
-test_that('VALUENOTE(X) is modified', {
+test_that('px_valuenote() and px_valuenotex()', {
   x <-
     'BEXSTA' %>%
     get_data_path() %>%
@@ -53,4 +53,54 @@ test_that('VALUENOTE(X) is modified', {
   x6 <- px_valuenotex(x_lang, valuenotex_df1)
 
   expect_identical(px_valuenotex(x6), valuenotex_df1)
+})
+
+test_that('px_values()', {
+  x <-
+    'BEXSTA' %>%
+    get_data_path() %>%
+    readRDS() %>%
+    px()
+
+  x_lang <- px_languages(x, c('dk', 'kl'))
+
+  x1 <-
+    x %>%
+    px_values(dplyr::tibble(`variable-code` = 'gender',
+                            code = 'K',
+                            value = 'Kvinde'
+                            )
+              )
+
+  expect <-
+    dplyr::tribble(
+     ~`variable-code`,  ~code,            ~language,   ~value,
+             "gender",    "K",        NA_character_, "Kvinde",
+             "gender",    "M",        NA_character_,      "M",
+             "gender",    "T",        NA_character_,      "T",
+     "place of birth",    "N",        NA_character_,      "N",
+     "place of birth",    "S",        NA_character_,      "S",
+     "place of birth",    "T",        NA_character_,      "T",
+               "time", "2018",        NA_character_,   "2018",
+               "time", "2019",        NA_character_,   "2019",
+               "time", "2020",        NA_character_,   "2020",
+               "time", "2021",        NA_character_,   "2021",
+               "time", "2022",        NA_character_,   "2022"
+                  )
+
+  expect_identical(px_values(x1), expect)
+
+  values_df <-
+    dplyr::tribble( ~`variable-code`,  ~code, ~language,           ~value,
+                              "time", "2018",      "dk",        "år 2018",
+                              "time", "2018",      "kl",  "ukiumi 2018-m"
+                    )
+
+  x2 <- x_lang %>% px_values(values_df)
+
+  expect_identical(px_values(x2), values_df)
+
+  x3 <- x2 %>% px_values(NULL)
+
+  expect_identical(px_values(x3), NULL)
 })